backend/requirements.txt
fastapi
uvicorn
ultralytics
pillow
python-multipart
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
psycopg2-binary
supabase
